refactor: 공통 참조가 필요한 util을 core 모듈로 이전
